Set Ruby version
We use the legacy `.ruby-version` file instead of `asdf`'s default file
`.tool-versions` because Rails will create a `.ruby-version` file for
us regardless and we don't want to be defining the Ruby version in
multiple places.

To enable `asdf`'s Ruby plugin to recognise the legacy config file we
need to update our `.asdfrc` file as documented in the plugin's
README[1]. Essentially we're adding

    legacy_version_file = yes

to the `asdf` config file.

---

1. https://github.com/asdf-vm/asdf-ruby#migrating-from-another-ruby-version-manager